Refactor tests to improve population logic and clarity
