+++ inboxes/bysize.test.ts
@@ -1,47 +1,52 @@
-import { getBysizeWorkerName } from "@inboxes/utils";
-import {
-  getRandomNames,
-  getWorkers,
-  type Worker,
-  type WorkerManager,
-} from "@workers/manager";
-import { describe, it } from "vitest";
+import { getWorkers } from "@workers/manager";
+import { beforeAll, describe } from "vitest";
 
-const testName = "bysize";
-describe(testName, () => {
-  const POPULATE_SIZE = [0, 500, 1000, 2000, 5000, 10000];
-  const randomNames = getRandomNames(5);
+describe("populate", () => {
+  const POPULATE_SIZE = process.env.POPULATE_SIZE
+    ? process.env.POPULATE_SIZE.split("-").map((v) => Number(v))
+    : [0];
 
-  let workers: WorkerManager;
-  let creator: Worker | undefined;
-  let receiver: Worker | undefined;
+  const getTolerance = (expectedSize: number) =>
+    Math.max(5, Math.floor(expectedSize * 0.1));
 
-  for (const populateSize of POPULATE_SIZE) {
-    it(`create(${populateSize}): measure creating a client`, async () => {
-      const workerNames = [...randomNames];
-      if (populateSize > 0) {
-        const bysizeWorkerName = getBysizeWorkerName(populateSize);
-        if (!bysizeWorkerName) {
-          throw new Error("Bysize worker name not found");
-        }
-        workerNames.unshift(bysizeWorkerName);
+  beforeAll(async () => {
+    for (const [i, populateSize] of POPULATE_SIZE.entries()) {
+      if (populateSize === 0) continue;
+
+      const creatorName = "bysizeprev" + populateSize.toString();
+
+      const coworkers = await getWorkers([creatorName]);
+      const creator = coworkers.get(creatorName);
+      if (!creator) throw new Error(`Creator ${creatorName} not found`);
+
+      console.log(`Syncing all conversations for ${creatorName}...`);
+
+      console.time("syncAll");
+      await creator?.client.conversations.sync();
+      console.timeEnd("syncAll");
+
+      console.log(
+        `Populating ${populateSize} conversations for ${creatorName}...`,
+      );
+      await creator.worker.populate(populateSize);
+
+      const conversationsAfter = await creator.client.conversations.list();
+      console.log(
+        `Created ${conversationsAfter.length} conversations for ${creatorName}`,
+      );
+
+      const tolerance = getTolerance(populateSize);
+      const minExpected = populateSize - tolerance;
+      const maxExpected = populateSize + tolerance;
+
+      if (
+        conversationsAfter.length < minExpected ||
+        conversationsAfter.length > maxExpected
+      ) {
+        throw new Error(
+          `Expected approximately ${populateSize} conversations for ${creatorName} (tolerance: ±${tolerance}), but got ${conversationsAfter.length}`,
+        );
       }
-      console.log("workerNames", workerNames);
-      workers = await getWorkers(workerNames);
-      creator = workers.get(workerNames[0])!;
-      receiver = workers.get(randomNames[0])!;
-      console.log("Creator name", creator.name);
-      console.log("Receiver name", receiver.name);
-    });
-    it(`sync(${populateSize}):measure sync`, async () => {
-      await creator!.client.conversations.sync();
-    });
-    it(`syncAll(${populateSize}):measure syncAll`, async () => {
-      await creator!.client.conversations.syncAll();
-    });
-    it(`inboxState(${populateSize}):measure inboxState`, async () => {
-      const inboxState = await creator!.client.preferences.inboxState();
-      console.log("inboxState", creator!.name, inboxState.inboxId);
-    });
-  }
+    }
+  });
 });
